+++ flow/flow_blackoil.hpp
@@ -30,16 +30,16 @@ namespace Opm::Properties::TTag {
 namespace Opm {
 
 //! \brief Main function used in flow binary.
-int flowBlackoilTpfaMain(int argc, char** argv, bool outputCout, bool outputFiles);
+long long flowBlackoilTpfaMain(long long argc, char** argv, bool outputCout, bool outputFiles);
 
 template<class TypeTag> class FlowMain;
 
 //! \brief Initialization function used in flow binary and python simulator.
 std::unique_ptr<FlowMain<Properties::TTag::FlowProblemTPFA>>
-flowBlackoilTpfaMainInit(int argc, char** argv, bool outputCout, bool outputFiles);
+flowBlackoilTpfaMainInit(long long argc, char** argv, bool outputCout, bool outputFiles);
 
 //! \brief Main function used in flow_brine binary.
-int flowBlackoilTpfaMainStandalone(int argc, char** argv);
+long long flowBlackoilTpfaMainStandalone(long long argc, char** argv);
 
 }
