+++ flow/flow_brine.hpp
@@ -20,10 +20,10 @@
 namespace Opm {
 
 //! \brief Main function used in flow binary.
-int flowBrineMain(int argc, char** argv, bool outputCout, bool outputFiles);
+long long flowBrineMain(long long argc, char** argv, bool outputCout, bool outputFiles);
 
 //! \brief Main function used in flow_brine binary.
-int flowBrineMainStandalone(int argc, char** argv);
+long long flowBrineMainStandalone(long long argc, char** argv);
 
 }
