+++ flow/flow_extbo.cpp
@@ -39,7 +39,7 @@ struct EnableExtbo<TypeTag, TTag::FlowExtboProblem> {
 namespace Opm {
 
 // ----------------- Main program -----------------
-int flowExtboMain(int argc, char** argv, bool outputCout, bool outputFiles)
+long long flowExtboMain(long long argc, char** argv, bool outputCout, bool outputFiles)
 {
     // we always want to use the default locale, and thus spare us the trouble
     // with incorrect locale settings.
@@ -50,7 +50,7 @@ int flowExtboMain(int argc, char** argv, bool outputCout, bool outputFiles)
     return mainfunc.execute();
 }
 
-int flowExtboMainStandalone(int argc, char** argv)
+long long flowExtboMainStandalone(long long argc, char** argv)
 {
     using TypeTag = Properties::TTag::FlowExtboProblem;
     auto mainObject = std::make_unique<Opm::Main>(argc, argv);
