+++ flow/flow_foam.hpp
@@ -20,10 +20,10 @@
 namespace Opm {
 
 //! \brief Main function used in flow binary.
-int flowFoamMain(int argc, char** argv, bool outputCout, bool outputFiles);
+long long flowFoamMain(long long argc, char** argv, bool outputCout, bool outputFiles);
 
 //! \brief Main function used in flow_foam binary.
-int flowFoamMainStandalone(int argc, char** argv);
+long long flowFoamMainStandalone(long long argc, char** argv);
 
 }
