+++ flow/flow_gasoil_energy.cpp
@@ -78,7 +78,7 @@ struct EnableDispersion<TypeTag, TTag::FlowGasOilEnergyProblem> { static constex
 namespace Opm {
 
 // ----------------- Main program -----------------
-int flowGasOilEnergyMain(int argc, char** argv, bool outputCout, bool outputFiles)
+long long flowGasOilEnergyMain(long long argc, char** argv, bool outputCout, bool outputFiles)
 {
     // we always want to use the default locale, and thus spare us the trouble
     // with incorrect locale settings.
@@ -89,7 +89,7 @@ int flowGasOilEnergyMain(int argc, char** argv, bool outputCout, bool outputFile
     return mainfunc.execute();
 }
 
-int flowGasOilEnergyMainStandalone(int argc, char** argv)
+long long flowGasOilEnergyMainStandalone(long long argc, char** argv)
 {
     using TypeTag = Properties::TTag::FlowGasOilEnergyProblem;
     auto mainObject = std::make_unique<Opm::Main>(argc, argv);
